requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via tutor
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   kubernetes
    #   requests
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via tutor
google-auth==2.22.0
    # via kubernetes
idna==3.4
    # via requests
jinja2==3.1.2
    # via tutor
kubernetes==27.2.0
    # via tutor
markupsafe==2.1.3
    # via jinja2
mypy==1.5.1
    # via tutor
mypy-extensions==1.0.0
    # via mypy
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycryptodome==3.18.0
    # via tutor
python-dateutil==2.8.2
    # via kubernetes
python-dotenv==1.0.0
    # via kubernetes
pyyaml==6.0.1
    # via
    #   kubernetes
    #   tutor
requests==2.31.0
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
tomli==2.0.1
    # via mypy
# tutor==16.1.8
    # via
    #   -r requirements.in
    #   tutor-cairn
    #   tutor-forum
    #   tutor-mfe
    #   tutor-minio
# tutor-cairn==16.0.3
    # via -r requirements.in
tutor-forum==16.0.2
    # via -r requirements.in
tutor-mfe==16.1.3
    # via -r requirements.in
# tutor-minio==16.0.3
    # via -r requirements.in
typing-extensions==4.7.1
    # via
    #   mypy
    #   tutor
urllib3==1.26.16
    # via
    #   google-auth
    #   kubernetes
    #   requests
websocket-client==1.6.2
    # via kubernetes
